Get rid of EffectCategory Enum
