+++ src/main/java/rip/shukaaa/effect/category/EffectCategoryRegister.java
@@ -1,0 +1,8 @@
+package rip.shukaaa.effect.category;
+
+public class EffectCategoryRegister {
+    public static final EffectCategory BASICS = new EffectCategory("Effects", "Basics");
+    public static final EffectCategory GENERATE = new EffectCategory("Effects", "Generate");
+    public static final EffectCategory DISTORTION = new EffectCategory("Effects", "Distortion");
+    public static final EffectCategory TRANSFORM = new EffectCategory("Image", "Transform");
+}
