+++ src/main/java/rip/shukaaa/effect/effects/basic/Melt.java
@@ -1,10 +1,10 @@
 package rip.shukaaa.effect.effects.basic;
 
-import rip.shukaaa.effect.Effect;
+import rip.shukaaa.effect.category.EffectCategoryRegister;
+import rip.shukaaa.effect.effects.Effect;
 import rip.shukaaa.effect.input.EffectInput;
 import rip.shukaaa.effect.input.inputs.ColorChooser;
 import rip.shukaaa.effect.input.inputs.Slider;
-import rip.shukaaa.enums.EffectCategory;
 import rip.shukaaa.exceptions.EffectOptionNotFoundException;
 import rip.shukaaa.image.Pixel;
 import rip.shukaaa.image.ShukaaaImage;
@@ -17,7 +17,7 @@ public class Melt extends Effect {
         super(new EffectInput[]{
                 new Slider(0, 256, 256, "Threshold"),
                 new ColorChooser("Color")
-        }, EffectCategory.EFFECTS);
+        }, EffectCategoryRegister.BASICS);
     }
 
     public ArrayList<Pixel> apply(ShukaaaImage image, HashMap<String, Object> args) throws EffectOptionNotFoundException {
