+++ src/main/java/rip/shukaaa/effect/effects/distortion/RandomShuffler.java
@@ -1,8 +1,8 @@
 package rip.shukaaa.effect.effects.distortion;
 
-import rip.shukaaa.effect.Effect;
+import rip.shukaaa.effect.category.EffectCategoryRegister;
+import rip.shukaaa.effect.effects.Effect;
 import rip.shukaaa.effect.input.EffectInput;
-import rip.shukaaa.enums.EffectCategory;
 import rip.shukaaa.exceptions.EffectOptionNotFoundException;
 import rip.shukaaa.image.Pixel;
 import rip.shukaaa.image.ShukaaaImage;
@@ -12,7 +12,7 @@ import java.util.HashMap;
 
 public final class RandomShuffler extends Effect {
     public RandomShuffler() {
-        super(new EffectInput[0], EffectCategory.DISTORTION);
+        super(new EffectInput[0], EffectCategoryRegister.DISTORTION);
     }
 
     public ArrayList<Pixel> apply(ShukaaaImage image, HashMap<String, Object> args) throws EffectOptionNotFoundException {
