+++ src/main/java/rip/shukaaa/effect/effects/distortion/TriangleRuler.java
@@ -1,9 +1,9 @@
 package rip.shukaaa.effect.effects.distortion;
 
-import rip.shukaaa.effect.Effect;
+import rip.shukaaa.effect.category.EffectCategoryRegister;
+import rip.shukaaa.effect.effects.Effect;
 import rip.shukaaa.effect.input.EffectInput;
 import rip.shukaaa.effect.input.inputs.Slider;
-import rip.shukaaa.enums.EffectCategory;
 import rip.shukaaa.exceptions.EffectOptionNotFoundException;
 import rip.shukaaa.image.Pixel;
 import rip.shukaaa.image.ShukaaaImage;
@@ -16,7 +16,7 @@ public class TriangleRuler extends Effect {
     public TriangleRuler() {
         super(new EffectInput[]{
                 new Slider(0, 100, 20, "Intensity"),
-        }, EffectCategory.DISTORTION);
+        }, EffectCategoryRegister.DISTORTION);
     }
 
     @Override
