+++ src/main/java/rip/shukaaa/effect/effects/transform/Degrees180Flip.java
@@ -1,8 +1,8 @@
 package rip.shukaaa.effect.effects.transform;
 
-import rip.shukaaa.effect.Effect;
+import rip.shukaaa.effect.category.EffectCategoryRegister;
+import rip.shukaaa.effect.effects.Effect;
 import rip.shukaaa.effect.input.EffectInput;
-import rip.shukaaa.enums.EffectCategory;
 import rip.shukaaa.exceptions.EffectOptionNotFoundException;
 import rip.shukaaa.image.Pixel;
 import rip.shukaaa.image.ShukaaaImage;
@@ -13,7 +13,7 @@ import java.util.HashMap;
 
 public class Degrees180Flip extends Effect {
     public Degrees180Flip() {
-        super(new EffectInput[0], EffectCategory.TRANSFORM);
+        super(new EffectInput[0], EffectCategoryRegister.TRANSFORM);
     }
 
     @Override
